Isolate tests for users to its own file
