+++ tests/api/index.js
@@ -33,57 +33,4 @@ var admin = nano.login({
   password: 'admin'
 });
 
-describe("List users", function() {
-
-  var expectedSchema = {
-    type: 'object',
-    properties: {
-      email: {type: 'string'},
-      activated: {type: 'boolean'},
-      is_admin: {type: 'boolean'},
-      first_name: {type: 'string'},
-      last_name: {type: 'string'},
-      sam: {type: 'string'},
-      windows_password: {type: 'string'},
-    },
-    required: ['email', 'first_name', 'activated', 'is_admin', 'first_name', 'last_name', 'sam', 'windows_password'],
-    additionalProperties: false
-  };
-
-  var request = nano.as(admin).get("api/users")
-      .shouldReturn(200)
-      .shouldBeJSONAPI()
-      .shouldComplyTo(expectedSchema);
-
-  it("should contain the admin",  function() {
-    return expect(request).to.comprise.of.json({
-      email: 'admin@nanocloud.com',
-      activated: true,
-      is_admin: true,
-      first_name: "John",
-      last_name: "Doe",
-      sam: "Administrator",
-      windows_password: "Nanocloud123+"
-    });
-  })
-})
-
-describe("Create user", function() {
-
-  var request = nano.as(admin).post("api/users", {
-    "data" : {
-      "type": "user",
-      "attributes": {
-        "first_name": "{{TEST_USER_FIRSTNAME}}",
-        "last_name": "{{TEST_USER_LASTNAME}}",
-        "email": "{{TEST_USER_EMAIL}}",
-        "password": "{{TEST_USER_PASSWORD}}"
-      }
-    }
-  }).shouldReturn(200)
-      .shouldBeJSON();
-
-  it("Should be valid JSON", function() {
-    expect(request).to.have.status(403);
-  })
-})
+require('./test-users')(admin);
